# [7.1.0] Starlark: reuse positional array in native calls where possible